ref: #30284

+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-binary-field/dot-edit-content-binary-field.component.ts
@@ -41,13 +41,13 @@ import {
     DotGeneratedAIImage
 } from '@dotcms/dotcms-models';
 import {
+    DotAIImagePromptComponent,
     DotDropZoneComponent,
     DotMessagePipe,
     DotSpinnerModule,
     DropZoneErrorType,
     DropZoneFileEvent,
-    DropZoneFileValidity,
-    DotAIImagePromptComponent
+    DropZoneFileValidity
 } from '@dotcms/ui';
 
 import { DotBinaryFieldEditorComponent } from './components/dot-binary-field-editor/dot-binary-field-editor.component';
@@ -76,7 +76,6 @@ type SystemOptionsType = {
 
 @Component({
     selector: 'dot-edit-content-binary-field',
-    standalone: true,
     imports: [
         CommonModule,
         ButtonModule,
@@ -144,6 +143,7 @@ export class DotEditContentBinaryFieldComponent
 
     $field = signal<DotCMSContentTypeField>({} as DotCMSContentTypeField);
     $variable = computed(() => this.$field()?.variable);
+    $disabled = signal<boolean>(false);
 
     @Output() valueUpdated = new EventEmitter<{ value: string; fileName: string }>();
     @ViewChild('inputFile') inputFile: ElementRef;
@@ -247,6 +247,10 @@ export class DotEditContentBinaryFieldComponent
         this.onTouched = fn;
     }
 
+    setDisabledState(isDisabled: boolean): void {
+        this.$disabled.set(isDisabled);
+    }
+
     ngOnDestroy() {
         this.#dotBinaryFieldEditImageService.removeListener();
         this.#dialogRef?.close();
@@ -259,6 +263,10 @@ export class DotEditContentBinaryFieldComponent
      * @memberof DotEditContentBinaryFieldComponent
      */
     openDialog(mode: BinaryFieldMode) {
+        if (this.$disabled()) {
+            return;
+        }
+
         if (mode === BinaryFieldMode.AI) {
             this.openAIImagePrompt();
         } else {
@@ -322,6 +330,10 @@ export class DotEditContentBinaryFieldComponent
      * @memberof DotEditContentBinaryFieldComponent
      */
     openFilePicker() {
+        if (this.$disabled()) {
+            return;
+        }
+
         this.inputFile.nativeElement.click();
     }
 
@@ -332,6 +344,10 @@ export class DotEditContentBinaryFieldComponent
      * @memberof DotEditContentBinaryFieldComponent
      */
     handleFileSelection(event: Event) {
+        if (this.$disabled()) {
+            return;
+        }
+
         const input = event.target as HTMLInputElement;
         const file = input.files[0];
         this.#dotBinaryFieldStore.handleUploadFile(file);
@@ -343,6 +359,10 @@ export class DotEditContentBinaryFieldComponent
      * @memberof DotEditContentBinaryFieldComponent
      */
     removeFile() {
+        if (this.$disabled()) {
+            return;
+        }
+
         this.#dotBinaryFieldStore.removeFile();
     }
 
@@ -397,6 +417,10 @@ export class DotEditContentBinaryFieldComponent
      * @memberof DotEditContentBinaryFieldComponent
      */
     handleFileDrop({ validity, file }: DropZoneFileEvent): void {
+        if (this.$disabled()) {
+            return;
+        }
+
         if (!validity.valid) {
             this.handleFileDropError(validity);
